feat: Support for RollingFileAppender date syntax
Adds support for naming the buffer files with the naming syntax supported by RollingFileAppender. This allows the files to be rotated on half hour and hour intervals, in addition to the exisiting daily rotation. Also adds support for the number of retained files.

fix: Removes trailing slash from bookmark

minor code and document changes